fixed incompatible types

+++ src/soundm4a.c
@@ -7,7 +7,7 @@ soundm4a.c
 
 */
 
-MA_API ma_result m4a_read_pcm_frames_wrapper(void *pDecoder, void *pFramesOut, long unsigned int frameCount, long unsigned int *pFramesRead)
+MA_API ma_result m4a_read_pcm_frames_wrapper(void *pDecoder, void *pFramesOut, size_t frameCount, size_t *pFramesRead)
 {
         ma_decoder *dec = (ma_decoder *)pDecoder;
         return m4a_decoder_read_pcm_frames((m4a_decoder *)dec->pUserData, pFramesOut, frameCount, (ma_uint64 *)pFramesRead);
